moved "getPeriodDateFormat" closer to the stats-query
